requirements.txt
gspread
google-auth
pandas
pyqt5
matplotlib
mplcursors